Module aws_sdk_appconfig::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

payload_too_large_exception

See PayloadTooLargeException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

Structs

BadRequestException

The input fails to satisfy the constraints specified by an AWS service.

ConflictException

The request could not be processed because of conflict in the current state of the

CreateApplicationError

CreateConfigurationProfileError

CreateDeploymentStrategyError

CreateEnvironmentError

CreateHostedConfigurationVersionError

DeleteApplicationError

DeleteConfigurationProfileError

DeleteDeploymentStrategyError

DeleteEnvironmentError

DeleteHostedConfigurationVersionError

GetApplicationError

GetConfigurationError

GetConfigurationProfileError

GetDeploymentError

GetDeploymentStrategyError

GetEnvironmentError

GetHostedConfigurationVersionError

InternalServerException

There was an internal failure in the AppConfig service.

ListApplicationsError

ListConfigurationProfilesError

ListDeploymentStrategiesError

ListDeploymentsError

ListEnvironmentsError

ListHostedConfigurationVersionsError

ListTagsForResourceError

PayloadTooLargeException

The configuration size is too large.

ResourceNotFoundException

The requested resource could not be found.

ServiceQuotaExceededException

The number of hosted configuration versions exceeds the limit for the AppConfig

StartDeploymentError

StopDeploymentError

TagResourceError

UntagResourceError

UpdateApplicationError

UpdateConfigurationProfileError

UpdateDeploymentStrategyError

UpdateEnvironmentError

ValidateConfigurationError

Enums

CreateApplicationErrorKind

CreateConfigurationProfileErrorKind

CreateDeploymentStrategyErrorKind

CreateEnvironmentErrorKind

CreateHostedConfigurationVersionErrorKind

DeleteApplicationErrorKind

DeleteConfigurationProfileErrorKind

DeleteDeploymentStrategyErrorKind

DeleteEnvironmentErrorKind

DeleteHostedConfigurationVersionErrorKind

GetApplicationErrorKind

GetConfigurationErrorKind

GetConfigurationProfileErrorKind

GetDeploymentErrorKind

GetDeploymentStrategyErrorKind

GetEnvironmentErrorKind

GetHostedConfigurationVersionErrorKind

ListApplicationsErrorKind

ListConfigurationProfilesErrorKind

ListDeploymentStrategiesErrorKind

ListDeploymentsErrorKind

ListEnvironmentsErrorKind

ListHostedConfigurationVersionsErrorKind

ListTagsForResourceErrorKind

StartDeploymentErrorKind

StopDeploymentErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApplicationErrorKind

UpdateConfigurationProfileErrorKind

UpdateDeploymentStrategyErrorKind

UpdateEnvironmentErrorKind

ValidateConfigurationErrorKind