Module aws_sdk_appconfig::operation

Structs

CreateApplication

An application in AppConfig is a logical unit of code that provides capabilities for your

CreateConfigurationProfile

Information that enables AppConfig to access the configuration source. Valid

CreateDeploymentStrategy

A deployment strategy defines important criteria for rolling out your configuration to

CreateEnvironment

For each application, you define one or more environments. An environment is a logical

CreateHostedConfigurationVersion

Create a new configuration in the AppConfig configuration store.

DeleteApplication

Delete an application. Deleting an application does not delete a configuration from a

DeleteConfigurationProfile

Delete a configuration profile. Deleting a configuration profile does not delete a

DeleteDeploymentStrategy

Delete a deployment strategy. Deleting a deployment strategy does not delete a

DeleteEnvironment

Delete an environment. Deleting an environment does not delete a configuration from a

DeleteHostedConfigurationVersion

Delete a version of a configuration from the AppConfig configuration store.

GetApplication

Retrieve information about an application.

GetConfiguration

Receive information about a configuration.

GetConfigurationProfile

Retrieve information about a configuration profile.

GetDeployment

Retrieve information about a configuration deployment.

GetDeploymentStrategy

Retrieve information about a deployment strategy. A deployment strategy defines

GetEnvironment

Retrieve information about an environment. An environment is a logical deployment group

GetHostedConfigurationVersion

Get information about a specific configuration version.

ListApplications

List all applications in your AWS account.

ListConfigurationProfiles

Lists the configuration profiles for an application.

ListDeploymentStrategies

List deployment strategies.

ListDeployments

Lists the deployments for an environment.

ListEnvironments

List the environments for an application.

ListHostedConfigurationVersions

View a list of configurations stored in the AppConfig configuration store by

ListTagsForResource

Retrieves the list of key-value tags assigned to the resource.

StartDeployment

Starts a deployment.

StopDeployment

Stops a deployment. This API action works only on deployments that have a status of

TagResource

Metadata to assign to an AppConfig resource. Tags help organize and categorize your

UntagResource

Deletes a tag key and value from an AppConfig resource.

UpdateApplication

Updates an application.

UpdateConfigurationProfile

Updates a configuration profile.

UpdateDeploymentStrategy

Updates a deployment strategy.

UpdateEnvironment

Updates an environment.

ValidateConfiguration

Uses the validators in a configuration profile to validate a configuration.